requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
discord.py==2.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosqlite==0.19.0
